% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cmd_j48}
\alias{cmd_j48}
\title{Create bash command for weka's J48}
\usage{
cmd_j48(C, M, data = "data/glass.arff",
  CLASSPATH = "export CLASSPATH=/Users/mbtangotan/Desktop/weka-3-7-11/weka.jar")
}
\arguments{
\item{C}{Cofidence Factor}

\item{M}{Amount of observations per leaf}

\item{data}{data/zaraza.arff}

\item{CLASSPATH}{export CLASSPATH=/pathtoweka/weka-3-7-11/weka.jar}
}
\description{
Create bash command for weka's J48
}

