% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{parse_j48}
\alias{parse_j48}
\title{Parse text of j48 output}
\usage{
parse_j48(model)
}
\arguments{
\item{model}{Output of `j48()`
## Not Run
## output <- j48(cmd)
## models <- parse_j48(output)}
}
\description{
Uses simple regex to extract the most valuable information from
the j48 classifier output.
}

